microservices/controller-service/requirements.txt
uvicorn
fastapi
python-jose
requests
python-multipart